+++ wms/wms_investment/doctype/wms_po_investment/wms_po_investment.py
@@ -2,6 +2,7 @@
 # For license information, please see license.txt
 
 import frappe
+from frappe import _
 from frappe.model.document import Document
 from frappe.model.naming import make_autoname
 from frappe.utils import add_months, nowdate
@@ -24,9 +25,11 @@ class WMSPOInvestment(Document):
 		account_number: DF.Data | None
 		amount: DF.Currency
 		client: DF.Link
+		client_classification: DF.Data | None
 		client_name: DF.Data | None
 		currency: DF.Link | None
 		entry_date: DF.Date
+		guardian: DF.Link | None
 		holders: DF.Table[WMSInvestmentHolder]
 		holding_type: DF.Link
 		is_active: DF.Check
@@ -67,6 +70,14 @@ class WMSPOInvestment(Document):
 		self.validate_nominee()
 		self.validate_dates()
 		self.validate_is_active()
+		self.validate_minor_investment()
+
+	def validate_minor_investment(self):
+		if self.client_classification == "Minor":
+			if self.guardian and self.guardian == self.client:
+				frappe.throw(_("Guardian cannot be the client."))
+			if self.holding_type != "Single":
+				frappe.throw(_("Holding Type must be 'Single' for Minor clients."))
 
 	def validate_is_active(self):
 		if self.status in ["Renewed", "Matured", "Pre-Matured", "Transmitted"]:
@@ -79,45 +90,47 @@ class WMSPOInvestment(Document):
 			for nominee in self.nominees:
 				share_percentage += nominee.share_percent
 				if nominee.nominee_name == self.client:
-					frappe.throw("Client cannot be a nominee.")
+					frappe.throw(_("Client cannot be a nominee."))
 				if nominee.nominee_name in holders:
-					frappe.throw("Holder cannot be a nominee.")
+					frappe.throw(_("Holder cannot be a nominee."))
+				if self.guardian and nominee.nominee_name == self.guardian:
+					frappe.throw(_("Guardian cannot be a nominee."))
 				if nominee.dob:
 					if calculate_age(nominee.dob) < 18:
 						nominee.is_minor = 1
 						if not nominee.guardian_name or not nominee.guardian_relation:
-							frappe.throw("Guardian details are required for minor nominees.")
+							frappe.throw(_("Guardian details are required for minor nominees."))
 						if nominee.guardian_dob and calculate_age(nominee.guardian_dob) < 18:
-							frappe.throw("Guardian must be at least 18 years old.")
+							frappe.throw(_("Guardian must be at least 18 years old."))
 						if nominee.guardian_name == self.client:
-							frappe.throw("Client cannot be a guardian.")
+							frappe.throw(_("Client cannot be a guardian."))
 						if nominee.nominee_name == nominee.guardian_name:
-							frappe.throw("Nominee and Guardian cannot be the same person.")
+							frappe.throw(_("Nominee and Guardian cannot be the same person."))
 						if nominee.guardian_name in holders:
-							frappe.throw("Holder cannot be the guardian.")
+							frappe.throw(_("Holder cannot be the guardian."))
 
 			if share_percentage != 100:
-				frappe.throw("Total share percentage of nominees must be 100%.")
+				frappe.throw(_("Total share percentage of nominees must be 100%."))
 
 	def validate_holders(self):
 		if self.holders and len(self.holders) > 0:
 			if self.holding_type == "Single":
-				frappe.throw("Cannot add holders for Single Holding Type.")
+				frappe.throw(_("Cannot add holders for Single Holding Type."))
 			if len(self.holders) > 2:
-				frappe.throw("You can only add a maximum of 2 holders.")
+				frappe.throw(_("You can only add a maximum of 2 holders."))
 			for holder in self.holders:
 				if holder.holder == self.client:
-					frappe.throw("Client cannot be a holder.")
+					frappe.throw(_("Client cannot be a holder."))
 
 	def validate_dates(self):
 		now_date = nowdate()
 
 		if self.entry_date and str(self.entry_date) > now_date:
-			frappe.throw("Entry Date cannot be in the future.")
+			frappe.throw(_("Entry Date cannot be in the future."))
 
 		if self.start_date:
 			if str(self.start_date) > now_date:
-				frappe.throw("Start Date cannot be in the future.")
+				frappe.throw(_("Start Date cannot be in the future."))
 			self.maturity_date = add_months(self.start_date, self.period)
 		else:
 			self.maturity_date = None